[fix] 서비스 핸들러 중복 코드 제거
