+++ src/domain/template/service/Comment.Service.ts
@@ -1,26 +1,11 @@
-import { CommentHandler } from './handler/CommentHandler';
+import { TemplateHandler } from './handler/TemplateHandler';
 import { Inject, Injectable } from '@nestjs/common';
-import { CommentOperation } from './types/comment';
+import { TemplateOperation } from './types/comment';
+import { BaseTemplateService } from './handler/BaseTemplate.Service';
 
 @Injectable()
-export class CommentService {
-  private handleMap = new Map<string, CommentHandler<any, any>>();
-  constructor(@Inject('COMMENT_HANDLERS') handlers: CommentHandler<any, any>[]) {
-    handlers.forEach((handler) => {
-      this.registerHandler(handler);
-    });
-  }
-
-  private registerHandler(handler: CommentHandler<any, any>) {
-    this.handleMap.set(handler.operation, handler);
-  }
-
-  async execute<Request extends unknown[] | unknown, Response>(
-    operation: CommentOperation,
-    ...request: Request extends unknown[] ? Request : [Request]
-  ): Promise<Response> {
-    const handler = this.handleMap.get(operation);
-    if (!handler) throw Error('Handler Empty Error');
-    return handler.handle(request) as Response;
+export class CommentService extends BaseTemplateService {
+  constructor(@Inject('COMMENT_HANDLERS') handlers: TemplateHandler<any, any>[]) {
+    super(handlers);
   }
 }
